BackEnd/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
supabase==2.0.2
python-dotenv==1.0.0
